Introduce EvaluationMetrics class
Add a class for all metrics. Corresponding members from TestUtil are moved into it.
